refactor: 프로필 이미지 url 추가 및 마이페이지 컨트롤러로 이동
