+++ src/main/java/balancetalk/member/application/MemberService.java
@@ -18,7 +18,6 @@ import balancetalk.member.domain.MemberRepository;
 import balancetalk.member.dto.ApiMember;
 import balancetalk.member.dto.MemberDto.JoinRequest;
 import balancetalk.member.dto.MemberDto.LoginRequest;
-import balancetalk.member.dto.MemberDto.MemberActivityResponse;
 import balancetalk.member.dto.MemberDto.MemberResponse;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -133,11 +132,6 @@ public class MemberService {
         if (memberRepository.existsByNickname(nickname)) {
             throw new BalanceTalkException(ALREADY_REGISTERED_NICKNAME);
         }
-    }
-
-    public MemberActivityResponse getMemberActivity(ApiMember apiMember) {
-        Member member = apiMember.toMember(memberRepository);
-        return MemberActivityResponse.fromEntity(member);
     }
 
     public String reissueAccessToken(HttpServletRequest request) {
